refactor(core): extract ANE compile cache and predict
Move compile_and_load plus the compiled-model disk cache into
bridge/cache.rs and Float16 predict_f32 into bridge/predict.rs.
Public bridge::compile_and_load / predict_f32 paths are unchanged.
